+++ controllers/cursosController.js
@@ -1,0 +1,10 @@
+const Curso = require('../models/Curso');
+
+exports.recuperarCursos = function (req, res) {
+  let cursos = new Curso();
+
+  cursos
+    .recuperarCursos()
+    .then((resultado) => console.log("cursos -- EXEMPLO"))
+    .catch((err) => res.send(err))
+}
